requirements.txt
opencv-python
keras
streamlit
streamlit-drawable-canvas
tensorflow
numpy
Pillow
